本地运行

+++ scripts/torch/register.py
@@ -52,8 +52,8 @@ model.to(device)
 model.eval()
 
 # set up tensors and permute
-input_moving = torch.from_numpy(moving).to(device).float().permute(0, 4, 1, 2, 3)
-input_fixed = torch.from_numpy(fixed).to(device).float().permute(0, 4, 1, 2, 3)
+input_moving = torch.from_numpy(moving.astype(np.float32)).to(device).float().permute(0, 4, 1, 2, 3)
+input_fixed = torch.from_numpy(fixed.astype(np.float32)).to(device).float().permute(0, 4, 1, 2, 3)
 
 # predict
 moved, warp = model(input_moving, input_fixed, registration=True)
